feat: remove Mnemonic verification
