+++ packages/dashboard/src/locales/en-US.json
@@ -38,7 +38,8 @@
     "setup_page_create_restore_button": "Recovery or Sign In",
     "create_account_identity_title": "Create an Identity for Mask Network",
     "create_account_sign_in_button": "Recovery",
-    "create_account_identity_warning": "Please don’t forget your identity code. Your identity code is the only proof of your user identity and the only basis for you to encrypt or decrypt social activities.",
+    "create_account_identity_warning": "Download your idendity ID to backup safely.",
+    "create_account_mnemonic_download_or_print": "I’ve downloaded it or printed it ",
     "create_account_mnemonic_confirm_failed": "Incorrect identity code",
     "create_account_connect_social_media_button": "Create",
     "create_account_connect_social_media": "Connect to {{type}}",
