+++ packages/dashboard/src/pages/SignUp/steps/PersonaCreate.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useState } from 'react'
 import { useLocation, useNavigate } from 'react-router-dom'
-import { MaskTextField, useCustomSnackbar } from '@masknet/theme'
+import { MaskTextField } from '@masknet/theme'
 import {
     Body,
     ColumnContentLayout,
@@ -8,15 +8,12 @@ import {
     PersonaLogoBox,
     SignUpAccountLogo,
 } from '../../../components/RegisterFrame/ColumnContentLayout'
-import { DashboardRoutes, delay } from '@masknet/shared-base'
+import { DashboardRoutes } from '@masknet/shared-base'
 import { Header } from '../../../components/RegisterFrame/ColumnContentHeader'
 import { Box, Typography } from '@mui/material'
 import { useDashboardI18N } from '../../../locales'
 import { SignUpRoutePath } from '../routePath'
-import { useCreatePersonaByPrivateKey, useCreatePersonaV2 } from '../../../hooks/useCreatePersonaV2'
-import { Services } from '../../../API'
 import { ButtonContainer } from '../../../components/RegisterFrame/ButtonContainer'
-import { PersonaContext } from '../../Personas/hooks/usePersonaContext'
 import { LoadingButton } from '../../../components/LoadingButton'
 
 const Label = ({ value }: { value: string }) => (
@@ -30,39 +27,17 @@ const Label = ({ value }: { value: string }) => (
 export const PersonaCreate = () => {
     const t = useDashboardI18N()
     const navigate = useNavigate()
-    const createPersona = useCreatePersonaV2()
-    const createPersonaByPrivateKey = useCreatePersonaByPrivateKey()
-    const { showSnackbar } = useCustomSnackbar()
-    const { changeCurrentPersona } = PersonaContext.useContainer()
-    const { state } = useLocation() as { state: { mnemonic: string[]; privateKey: string } }
 
     const [personaName, setPersonaName] = useState('')
-    const [error, setError] = useState('')
 
-    useEffect(() => {
-        if (
-            (!state?.mnemonic || !Services.Identity.validateMnemonic(state?.mnemonic.join(' '))) &&
-            !state?.privateKey
-        ) {
-            navigate(DashboardRoutes.SignUp, { replace: true })
-        }
-    }, [state?.mnemonic, state?.privateKey])
-
-    const create = useCallback(async () => {
-        try {
-            const identifier = state?.mnemonic
-                ? await createPersona(state?.mnemonic.join(' '), personaName)
-                : await createPersonaByPrivateKey(state?.privateKey, personaName)
-
-            await changeCurrentPersona(identifier)
-            showSnackbar(t.create_account_persona_successfully(), { variant: 'success' })
-
-            await delay(300)
-            navigate(`${DashboardRoutes.SignUp}/${SignUpRoutePath.ConnectSocialMedia}`)
-        } catch (error) {
-            setError((error as Error).message)
-        }
-    }, [state?.mnemonic, state?.privateKey, personaName])
+    const onNext = async () => {
+        await navigate(`${DashboardRoutes.SignUp}/${SignUpRoutePath.MnemonicReveal}`, {
+            replace: true,
+            state: {
+                personaName,
+            },
+        })
+    }
 
     return (
         <ColumnContentLayout>
@@ -82,15 +57,13 @@ export const PersonaCreate = () => {
                         InputProps={{ disableUnderline: true }}
                         onChange={(e) => setPersonaName(e.currentTarget.value)}
                         inputProps={{ maxLength: 24 }}
-                        error={!!error}
-                        helperText={error}
                     />
                     <ButtonContainer>
                         <LoadingButton
                             size="large"
                             variant="rounded"
                             color="primary"
-                            onClick={create}
+                            onClick={onNext}
                             disabled={!personaName}>
                             {t.next()}
                         </LoadingButton>
